[Refactor] Servicable AuthManager에서 분리 (#201)
